pyproject.toml
[tool.poetry]
name = "cha-de-panela"
version = "0.1.0"
description = "Cha de casa nova da Bea"
authors = ["joaovmvale <joao.marquesvale@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "^9.0.1"
flake8 = "^4.0.1"
djangorestframework = "^3.13.1"
django-environ = "^0.8.1"
django-heroku = "^0.3.1"
gunicorn = "^20.1.0"
whitenoise = "^6.0.0"
psycopg2 = "^2.9.3"
django-storages = "^1.12.3"
boto3 = "^1.21.9"

[tool.poetry.dev-dependencies]
black = {version = "^22.1.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
